Move db model & generated schema to appropriate module location
