+++ dsync-server/src/server.rs
@@ -2,14 +2,14 @@ use std::{process::Command, sync::Arc};
 
 use config::RunConfiguration;
 use database::DatabaseProxy;
-use diesel::{Connection, QueryDsl, RunQueryDsl, SelectableHelper, SqliteConnection};
+use diesel::{Connection, SqliteConnection};
 use dsync_proto::{
     cli::client_api_server::ClientApiServer, p2p::peer_service_server::PeerServiceServer,
 };
 use global_context::GlobalContext;
 use uuid::Uuid;
 
-use crate::models::LocalServerBaseInfoRow;
+use database::models::LocalServerBaseInfoRow;
 
 pub mod api;
 pub mod config;
