+++ dsync-server/src/server/database/models.rs
@@ -1,7 +1,7 @@
 use diesel::prelude::*;
 
 #[derive(Queryable, Selectable, Insertable)]
-#[diesel(table_name = crate::schema::peer_server_base_info)]
+#[diesel(table_name = super::schema::peer_server_base_info)]
 #[diesel(check_for_backend(diesel::sqlite::Sqlite))]
 pub struct PeerServerBaseInfoRow {
     pub uuid: String,
@@ -10,7 +10,7 @@ pub struct PeerServerBaseInfoRow {
 }
 
 #[derive(Queryable, Selectable, Insertable, Clone, Debug)]
-#[diesel(table_name = crate::schema::local_server_base_info)]
+#[diesel(table_name = super::schema::local_server_base_info)]
 #[diesel(check_for_backend(diesel::sqlite::Sqlite))]
 pub struct LocalServerBaseInfoRow {
     pub uuid: String,
@@ -19,7 +19,7 @@ pub struct LocalServerBaseInfoRow {
 }
 
 #[derive(Queryable, Selectable, Insertable, Clone)]
-#[diesel(table_name = crate::schema::peer_addr_v4)]
+#[diesel(table_name = super::schema::peer_addr_v4)]
 #[diesel(check_for_backend(diesel::sqlite::Sqlite))]
 pub struct PeerAddrV4Row {
     pub uuid: String,
